+++ extensions/dicom-segmentation/src/components/SegmentationPanel/SegmentationPanel.js
@@ -32,7 +32,11 @@ const refreshViewport = () => {
  * @param {Array} props.viewports - Viewports data (viewportSpecificData)
  * @param {number} props.activeIndex - Active viewport index
  * @param {boolean} props.isOpen - Boolean that indicates if the panel is expanded
- * @param {Function} props.onSegItemClick - Segment click handler
+ * @param {Function} props.onSegmentItemClick - Segment click handler
+ * @param {Function} props.onSegmentVisibilityChange - Segment visibiliy change handler
+ * @param {Function} props.onConfigurationChange - Configuration change handler
+ * @param {Function} props.activeContexts - List of active application contexts
+ * @param {Function} props.contexts - List of available application contexts
  * @returns component
  */
 const SegmentationPanel = ({
@@ -40,24 +44,40 @@ const SegmentationPanel = ({
   viewports,
   activeIndex,
   isOpen,
-  onSegItemClick,
+  onSegmentItemClick,
+  onSegmentVisibilityChange,
+  onConfigurationChange,
+  onDisplaySetLoadFailure,
+  onSelectedSegmentationChange,
+  activeContexts,
+  contexts,
 }) => {
+  const isVTK = () => activeContexts.includes(contexts.VTK);
+  const isCornerstone = () => activeContexts.includes(contexts.CORNERSTONE);
+
   /*
    * TODO: wrap get/set interactions with the cornerstoneTools
    * store with context to make these kind of things less blurry.
    */
   const { configuration } = cornerstoneTools.getModule('segmentation');
   const DEFAULT_BRUSH_RADIUS = configuration.radius || 10;
+
+  /*
+   * TODO: We shouldn't hardcode brushColor color, in the future
+   * the SEG may set the colorLUT to whatever it wants.
+   */
   const [state, setState] = useState({
     brushRadius: DEFAULT_BRUSH_RADIUS,
     brushColor:
-      'rgba(221, 85, 85, 1)' /* TODO: We shouldn't hardcode this color, in the future the SEG may set the colorLUT to whatever it wants. */,
+      'rgba(221, 85, 85, 1)',
     selectedSegment: null,
     selectedSegmentation: null,
-    showSegSettings: false,
+    showSegmentationSettings: false,
     brushStackState: null,
     labelmapList: [],
     segmentList: [],
+    cachedSegmentsProperties: [],
+    isLoading: false
   });
 
   useEffect(() => {
@@ -75,6 +95,25 @@ const SegmentationPanel = ({
     };
 
     /*
+     * TODO: Improve the way we notify parts of the app that depends on segs to be loaded.
+     *
+     * Currently we are using a non-ideal implementation through a custom event to notify the segmentation panel
+     * or other components that could rely on loaded segmentations that
+     * the segments were loaded so that e.g. when the user opens the panel
+     * before the segments are fully loaded, the panel can subscribe to this custom event
+     * and update itself with the new segments.
+     *
+     * This limitation is due to the fact that the cs segmentation module is an object (which will be
+     * updated after the segments are loaded) that React its not aware of its changes
+     * because the module object its not passed in to the panel component as prop but accessed externally.
+     *
+     * Improving this event approach to something reactive that can be tracked inside the react lifecycle,
+     * allows us to easily watch the module or the segmentations loading process in any other component
+     * without subscribing to external events.
+     */
+    document.addEventListener('extensiondicomsegmentationsegloaded', refreshSegmentations);
+
+    /*
      * These are specific to each element;
      * Need to iterate cornerstone-tools tracked enabled elements?
      * Then only care about the one tied to active viewport?
@@ -87,6 +126,7 @@ const SegmentationPanel = ({
     );
 
     return () => {
+      document.removeEventListener('extensiondicomsegmentationsegloaded', refreshSegmentations);
       cornerstoneTools.store.state.enabledElements.forEach(enabledElement =>
         enabledElement.removeEventListener(
           'cornerstonetoolslabelmapmodified',
@@ -94,9 +134,9 @@ const SegmentationPanel = ({
         )
       );
     };
-  });
+  }, [activeIndex, viewports]);
 
-  useEffect(() => {
+  const refreshSegmentations = useCallback(() => {
     const module = cornerstoneTools.getModule('segmentation');
     const activeViewport = viewports[activeIndex];
     const studyMetadata = studyMetadataManager.get(
@@ -106,7 +146,6 @@ const SegmentationPanel = ({
       activeViewport.displaySetInstanceUID
     );
     const brushStackState = module.state.series[firstImageId];
-
     if (brushStackState) {
       const labelmap3D =
         brushStackState.labelmaps3D[brushStackState.activeLabelmapIndex];
@@ -115,7 +154,11 @@ const SegmentationPanel = ({
         firstImageId,
         activeViewport
       );
-      const segmentList = getSegmentList(labelmap3D, firstImageId, brushStackState);
+      const segmentList = getSegmentList(
+        labelmap3D,
+        firstImageId,
+        brushStackState
+      );
       setState(state => ({
         ...state,
         brushStackState,
@@ -130,13 +173,21 @@ const SegmentationPanel = ({
         segmentList: [],
       }));
     }
-  }, [studies, viewports, activeIndex, getLabelmapList, getSegmentList, state.selectedSegmentation]);
+  }, [
+    viewports,
+    activeIndex,
+    state.isLoading
+  ]);
+
+  useEffect(() => {
+    refreshSegmentations();
+  }, [viewports, activeIndex, state.selectedSegmentation, activeContexts, state.isLoading]);
 
   /* Handle open/closed panel behaviour */
   useEffect(() => {
     setState(state => ({
       ...state,
-      showSegSettings: state.showSegSettings && !isOpen,
+      showSegmentationSettings: state.showSegmentationSettings && !isOpen,
     }));
   }, [isOpen]);
 
@@ -170,7 +221,9 @@ const SegmentationPanel = ({
               studies,
               displaySet,
               firstImageId,
-              brushStackState.activeLabelmapIndex
+              brushStackState.activeLabelmapIndex,
+              () => onSelectedSegmentationChange(),
+              onDisplaySetLoadFailure
             );
             updateState('selectedSegmentation', activatedLabelmapIndex);
           },
@@ -249,65 +302,104 @@ const SegmentationPanel = ({
               : prev;
           });
 
-          const enabledElements = cornerstone.getEnabledElements();
-          const element = enabledElements[activeIndex].element;
-          const toolState = cornerstoneTools.getToolState(element, 'stack');
+          if (isCornerstone()) {
+            const enabledElements = cornerstone.getEnabledElements();
+            const element = enabledElements[activeIndex].element;
+            const toolState = cornerstoneTools.getToolState(element, 'stack');
 
-          if (!toolState) {
-            return;
+            if (!toolState) {
+              return;
+            }
+
+            const imageIds = toolState.data[0].imageIds;
+            const imageId = imageIds[closest];
+            const frameIndex = imageIds.indexOf(imageId);
+
+            const SOPInstanceUID = cornerstone.metaData.get(
+              'SOPInstanceUID',
+              imageId
+            );
+            const StudyInstanceUID = cornerstone.metaData.get(
+              'StudyInstanceUID',
+              imageId
+            );
+
+            onSegmentItemClick({
+              StudyInstanceUID,
+              SOPInstanceUID,
+              frameIndex,
+              activeViewportIndex: activeIndex,
+            });
           }
 
-          const imageIds = toolState.data[0].imageIds;
-          const imageId = imageIds[closest];
-          const frameIndex = imageIds.indexOf(imageId);
+          if (isVTK()) {
+            const activeViewport = viewports[activeIndex];
+            const studyMetadata = studyMetadataManager.get(
+              activeViewport.StudyInstanceUID
+            );
+            const allDisplaySets = studyMetadata.getDisplaySets();
+            const currentDisplaySet = allDisplaySets.find(
+              displaySet =>
+                displaySet.displaySetInstanceUID ===
+                activeViewport.displaySetInstanceUID
+            );
 
-          const SOPInstanceUID = cornerstone.metaData.get(
-            'SOPInstanceUID',
-            imageId
-          );
-          const StudyInstanceUID = cornerstone.metaData.get(
-            'StudyInstanceUID',
-            imageId
-          );
+            const frame = labelmap3D.labelmaps2D[closest];
 
-          onSegItemClick({
-            StudyInstanceUID,
-            SOPInstanceUID,
-            frameIndex,
-            activeViewportIndex: activeIndex,
-          });
+            onSegmentItemClick({
+              studies,
+              StudyInstanceUID: currentDisplaySet.StudyInstanceUID,
+              displaySetInstanceUID: currentDisplaySet.displaySetInstanceUID,
+              SOPClassUID: viewports[activeIndex].sopClassUIDs[0],
+              SOPInstanceUID: currentDisplaySet.SOPInstanceUID,
+              segmentNumber,
+              frameIndex: closest,
+              frame,
+            });
+          }
         };
 
-        const enabledElements = cornerstone.getEnabledElements();
-        const enabledElementViewport = enabledElements[activeIndex];
+        const isSegmentVisible = () => {
+          return !labelmap3D.segmentsHidden[segmentIndex];
+        };
 
-        let isVisible = true;
-        if (enabledElementViewport) {
-          const element = enabledElementViewport.element;
-          const module = cornerstoneTools.getModule('segmentation');
-          isVisible = module.getters.isSegmentVisible(
-            element,
-            segmentNumber,
-            brushStackState.activeLabelmapIndex
-          );
+        const toggleSegmentVisibility = () => {
+          const segmentsHidden = labelmap3D.segmentsHidden;
+          segmentsHidden[segmentIndex] = !segmentsHidden[segmentIndex];
+          return !segmentsHidden[segmentIndex];
+        };
+
+        const cachedSegmentProperties = state.cachedSegmentsProperties[segmentNumber];
+        let visible = isSegmentVisible();
+        if (cachedSegmentProperties && cachedSegmentProperties.visible !== visible) {
+          toggleSegmentVisibility();
         }
 
         segmentList.push(
           <SegmentItem
             key={segmentNumber}
             itemClass={`segment-item ${sameSegment && 'selected'}`}
-            onClick={setCurrentSelectedSegment}
+            onClick={() => setCurrentSelectedSegment()}
             label={segmentLabel}
             index={segmentNumber}
             color={color}
-            visible={isVisible}
-            onVisibilityChange={() => {
-              const element = enabledElements[activeIndex].element;
-              module.setters.toggleSegmentVisibility(
-                element,
-                segmentNumber,
-                brushStackState.activeLabelmapIndex
-              );
+            visible={visible}
+            onVisibilityChange={newVisibility => {
+              if (isCornerstone()) {
+                const enabledElements = cornerstone.getEnabledElements();
+                const element = enabledElements[activeIndex].element;
+                module.setters.toggleSegmentVisibility(
+                  element,
+                  segmentNumber,
+                  brushStackState.activeLabelmapIndex
+                );
+              }
+
+              if (isVTK()) {
+                onSegmentVisibilityChange(segmentNumber, newVisibility);
+              }
+
+              updateCachedSegmentsProperties(segmentNumber, { visible: newVisibility });
               refreshViewport();
             }}
           />
@@ -323,8 +415,24 @@ const SegmentationPanel = ({
        * Show default name
        */
     },
-    [activeIndex, onSegItemClick, state.selectedSegment]
+    [activeIndex, onSegmentItemClick, state.selectedSegment, state.isLoading]
   );
+
+  const updateCachedSegmentsProperties = (segmentNumber, properties) => {
+    const segmentsProperties = state.cachedSegmentsProperties;
+    const segmentProperties = state.cachedSegmentsProperties[segmentNumber];
+
+    segmentsProperties[segmentNumber] =
+      segmentProperties ?
+        { ...segmentProperties, ...properties } :
+        properties;
+
+    updateState('cachedSegmentsProperties', segmentsProperties);
+  };
+
+  useEffect(() => {
+    updateState('cachedSegmentsProperties', []);
+  }, [activeContexts]);
 
   const updateState = (field, value) => {
     setState(state => ({ ...state, [field]: value }));
@@ -374,7 +482,6 @@ const SegmentationPanel = ({
   };
 
   const updateConfiguration = newConfiguration => {
-    /* Supported configuration */
     configuration.renderFill = newConfiguration.renderFill;
     configuration.renderOutline = newConfiguration.renderOutline;
     configuration.shouldRenderInactiveLabelmaps =
@@ -384,14 +491,17 @@ const SegmentationPanel = ({
     configuration.outlineWidth = newConfiguration.outlineWidth;
     configuration.fillAlphaInactive = newConfiguration.fillAlphaInactive;
     configuration.outlineAlphaInactive = newConfiguration.outlineAlphaInactive;
+    onConfigurationChange(newConfiguration);
     refreshViewport();
   };
 
-  if (state.showSegSettings) {
+  const disabledConfigurationFields = ['outlineAlpha', 'shouldRenderInactiveLabelmaps'];
+  if (state.showSegmentationSettings) {
     return (
       <SegmentationSettings
+        disabledFields={isVTK() ? disabledConfigurationFields : []}
         configuration={configuration}
-        onBack={() => updateState('showSegSettings', false)}
+        onBack={() => updateState('showSegmentationSettings', false)}
         onChange={updateConfiguration}
       />
     );
@@ -403,7 +513,7 @@ const SegmentationPanel = ({
           name="cog"
           width="25px"
           height="25px"
-          onClick={() => updateState('showSegSettings', true)}
+          onClick={() => updateState('showSegmentationSettings', true)}
         />
         {false && (
           <form className="selector-form">
@@ -424,7 +534,9 @@ const SegmentationPanel = ({
         <h3>Segmentations</h3>
         <div className="segmentations">
           <SegmentationSelect
-            value={state.labelmapList.find(i => i.value === state.selectedSegmentation)}
+            value={state.labelmapList.find(
+              i => i.value === state.selectedSegmentation
+            )}
             formatOptionLabel={SegmentationItem}
             options={state.labelmapList}
           />
@@ -509,7 +621,9 @@ const _setActiveLabelmap = async (
   studies,
   displaySet,
   firstImageId,
-  activeLabelmapIndex
+  activeLabelmapIndex,
+  callback = () => { },
+  onDisplaySetLoadFailure
 ) => {
   if (displaySet.labelmapIndex === activeLabelmapIndex) {
     log.warn(`${activeLabelmapIndex} is already the active labelmap`);
@@ -519,7 +633,17 @@ const _setActiveLabelmap = async (
   if (!displaySet.isLoaded) {
     // What props does this expect `viewportSpecificData` to have?
     // TODO: Should this return the `labelmapIndex`?
-    await displaySet.load(viewportSpecificData, studies);
+
+    const loadPromise = displaySet.load(viewportSpecificData, studies);
+
+    loadPromise.catch(error => {
+      onDisplaySetLoadFailure(error);
+
+      // Return old index.
+      return activeLabelmapIndex;
+    });
+
+    await loadPromise;
   }
 
   const { state } = cornerstoneTools.getModule('segmentation');
@@ -527,6 +651,8 @@ const _setActiveLabelmap = async (
   brushStackState.activeLabelmapIndex = displaySet.labelmapIndex;
 
   refreshViewport();
+
+  callback();
 
   return displaySet.labelmapIndex;
 };